testing utf-8 for source

+++ ML-Research/VectorCalculations.cpp
@@ -2,7 +2,7 @@
 
 int maxIndex(Eigen::VectorXd vec) {
 	assert(vec.size() > 0);
-	int max = vec[0];
+	double max = vec[0];
 	int ret = 0;
 	for (int i = 0; i < vec.size(); i++) {
 		if (vec[i] > max) {
